Agents-practice-june-2025/hello-agent-async/pyproject.toml
[project]
name = "hello-agent-async"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "wajidminhas", email = "shanitent667@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "2to3>=1.0",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]

[project.scripts]
hello-agent-async = "hello_agent_async:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
